requirements.txt
Flask==1.1.1
Flask-Login==0.4.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
PyMySQL==0.9.3
Flask-Misaka==1.0.0
Flask-Cors==3.0.8